Make ap_proto a regualar module
