+++ python/lsst/l1dbproto/__init__.py
@@ -1,3 +1,4 @@
 
 from .version import *
 from .config import *
+from .ap_proto import *
